requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argbind==0.3.9
asttokens==3.0.0
audiomentations==0.41.0
audioread==3.0.1
audioseal==0.1.7
bitarray==3.4.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
DeepFilterLib==0.5.6
DeepFilterNet==0.5.6
descript-audio-codec==1.0.0
descript-audiotools @ git+https://github.com/descriptinc/audiotools@348ebf2034ce24e2a91a553e3171cb00c0c71678
docstring_parser==0.16
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.58.1
fsspec==2025.3.2
future==1.0.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
grpcio==1.60.1
hf-xet==1.1.2
huggingface-hub==0.32.2
hydra-core==1.3.2
idna==3.10
importlib_resources==6.5.2
ipython==8.36.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
julius==0.2.7
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.5.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
noisereduce==3.0.3
numba==0.61.2
numpy==1.26.4
numpy-minmax==0.4.0
numpy-rms==0.5.0
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opuspy==0.0.3
packaging==23.2
pandas==2.3.0
parso==0.8.4
pedalboard==0.9.17
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==4.25.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pyloudnorm==0.1.1
pyparsing==3.2.3
pystoi==0.4.1
python-dateutil==2.8.2
python-stretch==0.3.1
pytz==2024.1
PyYAML==6.0.2
randomname==0.2.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rich==14.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torch-stoi==0.2.3
torch-tb-profiler==0.4.3
torchaudio==2.7.0
torchtyping==0.1.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
triton==3.3.0
typeguard==2.13.3
typing_extensions==4.13.2
tzdata==2023.4
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
